# api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CNSTRCT API MCP Server",
    "description": "CNSTRCT API for integrating with QuickBooks Online for transactions, expenses, and invoices",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/qbo/callback": {
      "get": {
        "summary": "OAuth callback",
        "description": "Callback endpoint for OAuth2 flow",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "Authorization code",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "State parameter",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "realmId",
            "in": "query",
            "description": "QuickBooks Company ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth token response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "token": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/qbo/retrieveToken": {
      "get": {
        "summary": "Retrieve current OAuth token",
        "description": "Returns the current OAuth token",
        "responses": {
          "200": {
            "description": "OAuth token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "No token found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/qbo/refreshAccessToken": {
      "get": {
        "summary": "Refresh OAuth access token",
        "description": "Refreshes the OAuth access token",
        "responses": {
          "200": {
            "description": "Refreshed OAuth token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}